splite webpack config #9
