+++ .env.dist
@@ -1,0 +1,2 @@
+HOST="localhost"
+DEVTOOLS=true
