--- env.json.dist
@@ -1,4 +1,0 @@
-{
-	"host": "localhost",
-	"devtools": true
-}
